+++ docs/.vuepress/config.js
@@ -79,6 +79,7 @@ module.exports = {
           title: 'prefect.tasks',
           collapsable: true,
           children: [
+            'tasks/aws',
             'tasks/control_flow',
             'tasks/function',
             'tasks/google',